+++ packages/ripple-ui-core/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0-alpha.54](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.53...v2.1.0-alpha.54) (2023-01-03)
+
+
+### Bug Fixes
+
+* **@dpc-sdp/ripple-ui-core:** page links disappeared from storybook ([57de409](https://github.com/dpc-sdp/ripple-framework/commit/57de4090f0e521e4cfb336ce2a0f7f3b5ebe14a8))
+* **@dpc-sdp/ripple-ui-core:** remove tabindex, update button to expose link, remove space ([a7e2540](https://github.com/dpc-sdp/ripple-framework/commit/a7e2540336ca9f418c65046315019c5c12a1071d))
+
+
+
+
+
+# [2.1.0-alpha.53](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.52...v2.1.0-alpha.53) (2022-12-30)
+
+
+### Bug Fixes
+
+* **@dpc-sdp/ripple-ui-core:** :lipstick: spacing tweaks for header ([0802720](https://github.com/dpc-sdp/ripple-framework/commit/0802720785868198764cefbcc9343e18354364dd))
+* **@dpc-sdp/ripple-ui-core:** :lipstick: style tweaks for cta secondary link ([334e88b](https://github.com/dpc-sdp/ripple-framework/commit/334e88bb2516c13a901bdf56c698d6c8bcc7dd08))
+* **@dpc-sdp/ripple-ui-core:** :lipstick: styling tweaks for header ([b8ef396](https://github.com/dpc-sdp/ripple-framework/commit/b8ef39683237463bcfe46ffa9eb791f64c13c2f2))
+
+
+### Features
+
+* **@dpc-sdp/ripple-ui-core:** :lipstick: change icon name input to dropdown rather than text box ([9a18b29](https://github.com/dpc-sdp/ripple-framework/commit/9a18b29677edcc0f57e264cf6ed510884e49aa96))
+
+
+
+
+
+# [2.1.0-alpha.52](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.51...v2.1.0-alpha.52) (2022-12-22)
+
+
+### Features
+
+* :sparkles: add neutral theme for header via feature flag ([dff7865](https://github.com/dpc-sdp/ripple-framework/commit/dff786541890ff8cfee0670b56f6f8ef7e289825))
+* **@dpc-sdp/ripple-tide-api:** :sparkles: adds site feature flags and sets footer theme variant ([05f8cb0](https://github.com/dpc-sdp/ripple-framework/commit/05f8cb03ae892960a427afa2c68f21b9a18c0ddc))
+* **@dpc-sdp/ripple-ui-core:** add neutral button theme ([94fcf74](https://github.com/dpc-sdp/ripple-framework/commit/94fcf742778d06f05dc05b694e608752e2d5c4e5))
+
+
+
+
+
+# [2.1.0-alpha.51](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.50...v2.1.0-alpha.51) (2022-12-22)
+
+
+### Features
+
+* **@dpc-sdp/ripple-ui-core:** add media embed to exports, use the new media embed in media template ([2241f1a](https://github.com/dpc-sdp/ripple-framework/commit/2241f1a926d12a52443de2ede32ba70e2b591201))
+* **@dpc-sdp/ripple-ui-core:** use list component for grant details ([d608542](https://github.com/dpc-sdp/ripple-framework/commit/d608542340e3ad031fd7fd86c56b5f12e444f2e0))
+
+
+
+
+
 # [2.1.0-alpha.50](https://github.com/dpc-sdp/ripple-framework/compare/v2.1.0-alpha.49...v2.1.0-alpha.50) (2022-12-21)
 
 **Note:** Version bump only for package @dpc-sdp/ripple-ui-core